requirements.txt
Django==5.1.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.9
Django==5.1.6
filelock==3.17.0
identify==2.6.6
nodeenv==1.9.1
platformdirs==4.3.6
PyYAML==6.0.2
sqlparse==0.5.3
virtualenv==20.29.1
